requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
googletrans==4.0.0rc1
pandas==2.1.3
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.13.3
beautifulsoup4==4.12.2
lxml==4.9.3
jinja2==3.1.2 